requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
requests>=2.31.0
feedparser>=6.0.0
pillow>=10.0.0
pytz>=2024.1
